Added new comparer for slicer items.
Moved some classes to its own files.
