+++ src/EPPlus/FormulaParsing/Excel/Functions/RefAndLookup/LookupUtils/LookupComparerBase.cs
@@ -40,42 +40,9 @@ namespace OfficeOpenXml.FormulaParsing.Excel.Functions.RefAndLookup.LookupUtils
             }
             else
             {
-                v = CompareObjects(x, y);
+                v = ComparerUtil.CompareObjects(x, y);
             }
             return v * (sortOrder > -1 ? 1 : -1);
-        }
-
-        protected int CompareObjects(object x1, object y1)
-        {
-            int ret;
-            var isNumX = ConvertUtil.IsNumericOrDate(x1);
-            var isNumY = ConvertUtil.IsNumericOrDate(y1);
-            if (isNumX && isNumY)   //Numeric Compare
-            {
-                var d1 = ConvertUtil.GetValueDouble(x1);
-                var d2 = ConvertUtil.GetValueDouble(y1);
-                if (double.IsNaN(d1))
-                {
-                    d1 = double.MaxValue;
-                }
-                if (double.IsNaN(d2))
-                {
-                    d2 = double.MaxValue;
-                }
-                ret = d1 < d2 ? -1 : d1 > d2 ? 1 : 0;
-            }
-            else if (isNumX == false && isNumY == false)   //String Compare
-            {
-                var s1 = x1 == null ? "" : x1.ToString();
-                var s2 = y1 == null ? "" : y1.ToString();
-                ret = string.Compare(s1, s2, StringComparison.CurrentCultureIgnoreCase);
-            }
-            else
-            {
-                ret = isNumX ? -1 : 1;
-            }
-
-            return ret;
         }
     }
 }
